+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support to search for a holding by `isin`, `name` and `symbol` (experimental)
 - Added support for notes in the activities import
 - Added support to search in the platform selector of the create or update account dialog
 - Added support for a search query in the portfolio position endpoint